# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Task:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample self hosted todo list api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo-host API
  version: "1.0"
paths:
  /task:
    post:
      consumes:
      - application/json
      description: Create a new task with the input payload
      parameters:
      - description: Create Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/entity.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
      security:
      - BasicAuth: []
      summary: Create a new task
      tags:
      - tasks
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by task id
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/entity.Task'
      security:
      - BasicAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a task by it id
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
      security:
      - BasicAuth: []
      summary: Get a task
      tags:
      - tasks
  /tasks:
    get:
      consumes:
      - application/json
      description: Get list of all current user tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
      security:
      - BasicAuth: []
      summary: Get task list
      tags:
      - tasks
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"